Moved code from acShared.h to rocketFSM.h
