+++ flight/mcu/lib/teensyShared/dataLog.h
@@ -11,7 +11,8 @@
 #include "SparkFun_Qwiic_KX13X.h"  //High-G IMU Library
 #include "SparkFun_u-blox_GNSS_Arduino_Library.h"
 #include "VoltageSensor.h"
-#include "acShared.h"
+#include "rocketFSM.h"
+// #include "acShared.h"
 // #include "dataStructs.h"
 
 /**
