+++ flight/mcu/lib/teensyShared/kalmanFilter.h
@@ -1,6 +1,6 @@
 #include "../EigenArduino-Eigen30/Eigen30.h"
 #include "ServoControl.h"
-#include "acShared.h"
+#include "rocketFSM.h"
 #include "dataLog.h"
 #include "rk4.h"
 #include "sensors.h"
