+++ flight/mcu/lib/teensyShared/rocketFSM.cpp
@@ -24,7 +24,7 @@
 #include <SPI.h>
 #include <Wire.h>
 
-#include "acShared.h"
+// #include "acShared.h"
 #include "dataLog.h"
 #include "hybridShared.h"
 #include "pins.h"
